+++ changelog.md
@@ -12,6 +12,10 @@
 - react-native-bpk-component-spinner: 0.0.1 => 1.0.0
   - New component `BpkSpinner`. See http://backpack.prod.aws.skyscnr.com/components/native/spinners
 
+**Fixed:**
+- bpk-component-popover:
+  - Fix for target element refocus after close
+
 ## 2017-10-31 - Fix for vertical alignment of text in select inputs
 
 **Fixed:**